enemy def

+++ Assets/UrbiEtUrbi/Scripts/Runtime/Helpers/IHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public interface IHealth 
 {
-    public void SetInitialHealth(int amount);
+    public void SetInitialHealth(float amount);
     public void ChangeHealth(float amount);
-    public void ChangeHealth(int amount, AttackType type);
+    public void ChangeHealth(float amount, AttackType type);
     public void Die();
 }
